Added two Models App and DAL. DAL will be deleted. App holds the information for my applications. Title, picture,Githublink... This will help in the future all the app list grows. Much like the home page skills.
I linked the app model to the controller and created all my app information here.
Passed it to my index

Since all of my app data is housed in the controller I didn't need all of the divs and what not.
So I deleted all but one and used it as a template.
This way I'm able to create applications links and such on the fly instead of by hand.
This will make it so much easier in the future with a big list of apps.